requirements.txt
torch==2.2.1
ultralytics==8.0.186
streamlit
pillow==10.2.0
opencv-python-headless